마켓/메인 - 마켓메인 /앱메인 관련 작업 적용

+++ app/src/main/java/com/Whowant/Tokki/UI/Activity/Market/MarketDetailActivity.java
@@ -81,8 +81,8 @@ public class MarketDetailActivity extends AppCompatActivity {
           carrotNum.text = String(Int( nPrice! / 120)) + "개"
 
          */
-        int nPrice = market.getPrice();
-        int nCarrot = nPrice / 120;
+        long nPrice = market.getPrice();
+        int nCarrot = (int)nPrice / 120;
 
         DecimalFormat formatter = new DecimalFormat("#,###,###");
         String formattedPrice = formatter.format(market.getPrice());
